fixes paths
