+++ testing/package.json
@@ -8,6 +8,7 @@
     "install-browsers": "playwright install chromium",
     "test:ui": "playwright test --ui",
     "test:debug": "playwright test --debug",
+    "test:api": "playwright test e2e/api-docs.spec.ts",
     "test:presentations-list": "playwright test e2e/presentations-list.spec.ts",
     "test:create-presentation": "playwright test e2e/create-presentation.spec.ts",
     "test:steps": "playwright test e2e/presentation-steps.spec.ts",
